requirements.txt
# NetSentinel Requirements

# Database
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.26.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Utilities
python-dateutil>=2.8.2